+++ packages/eyes-sdk-core/lib/server/ServerConnector.js
@@ -165,7 +165,6 @@ class ServerConnector {
     const validStatusCodes = [HTTP_STATUS_CODES.OK, HTTP_STATUS_CODES.CREATED]
     if (validStatusCodes.includes(response.status)) {
       const runningSession = new RunningSession(response.data)
-      runningSession.setNewSession(response.status === HTTP_STATUS_CODES.CREATED)
       this._logger.verbose('ServerConnector.startSession - post succeeded', runningSession)
       return runningSession
     }
@@ -192,7 +191,6 @@ class ServerConnector {
 
     const config = {
       name: 'stopSession',
-      isLongRequest: true,
       method: 'DELETE',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),
@@ -328,7 +326,6 @@ class ServerConnector {
 
     const config = {
       name: 'matchWindow',
-      isLongRequest: true,
       method: 'POST',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),
@@ -373,7 +370,6 @@ class ServerConnector {
 
     const config = {
       name: 'matchSingleWindow',
-      isLongRequest: true,
       method: 'POST',
       url: GeneralUtils.urlConcat(this._configuration.getServerUrl(), EYES_API_PATH),
       headers: {},
@@ -420,7 +416,6 @@ class ServerConnector {
 
     const config = {
       name: 'replaceWindow',
-      isLongRequest: true,
       method: 'PUT',
       url: GeneralUtils.urlConcat(
         this._configuration.getServerUrl(),
@@ -694,33 +689,33 @@ class ServerConnector {
    * @param {string} domJson
    * @return {Promise<string>}
    */
-  async postDomSnapshot(domJson) {
+  async postDomSnapshot(id, domJson) {
     ArgumentGuard.notNull(domJson, 'domJson')
     this._logger.verbose('ServerConnector.postDomSnapshot called')
+    const url = this._renderingInfo.getResultsUrl().replace('__random__', id)
 
     const config = {
       name: 'postDomSnapshot',
-      method: 'POST',
-      url: GeneralUtils.urlConcat(
-        this._configuration.getServerUrl(),
-        EYES_API_PATH,
-        '/running/data',
-      ),
+      retry: 3,
+      method: 'PUT',
+      url,
+      data: zlib.gzipSync(Buffer.from(domJson)),
       headers: {
+        Date: new Date().toISOString(),
+        'x-ms-blob-type': 'BlockBlob',
         'Content-Type': 'application/octet-stream',
       },
     }
 
-    config.data = zlib.gzipSync(Buffer.from(domJson))
-
     const response = await this._axios.request(config)
-    const validStatusCodes = [HTTP_STATUS_CODES.OK, HTTP_STATUS_CODES.CREATED]
-    if (validStatusCodes.includes(response.status)) {
-      this._logger.verbose('ServerConnector.postDomSnapshot - post succeeded')
-      return response.headers.location
+    if (response.status !== HTTP_STATUS_CODES.CREATED) {
+      throw new Error(
+        `ServerConnector.postDomSnapshot - unexpected status (${response.statusText})`,
+      )
     }
 
-    throw new Error(`ServerConnector.postDomSnapshot - unexpected status (${response.statusText})`)
+    this._logger.verbose('ServerConnector.postDomSnapshot - post succeeded')
+    return url
   }
 
   async getUserAgents() {