+++ packages/eyes-sdk-core/test/unit/server/ServerConnector.spec.js
@@ -30,7 +30,7 @@ describe('ServerConnector', () => {
         baselineId: `${sessionId}__baseline`,
         batchId,
         id: `${sessionId}__running`,
-        isNewSession: false,
+        isNew: undefined,
         renderingInfo: undefined,
         sessionId,
         url: `${sessionId}__url`,
@@ -83,13 +83,23 @@ describe('ServerConnector', () => {
     }
   })
 
-  it('uploadScreenshot works', async () => {
+  it('uploadScreenshot uploads to resultsUrl webhook', async () => {
     const configuration = new Configuration()
     const serverConnector = new ServerConnector(logger, configuration)
     const renderingInfo = await serverConnector.renderInfo()
     const id = GeneralUtils.guid()
     const buffer = Buffer.from('something')
     const result = await serverConnector.uploadScreenshot(id, buffer)
+    assert.strictEqual(result, renderingInfo.getResultsUrl().replace('__random__', id))
+  })
+
+  it('postDomSnapshot uploads to resultsUrl webhook', async () => {
+    const configuration = new Configuration()
+    const serverConnector = new ServerConnector(logger, configuration)
+    const renderingInfo = await serverConnector.renderInfo()
+    const buffer = Buffer.from('something')
+    const id = GeneralUtils.guid()
+    const result = await serverConnector.postDomSnapshot(id, buffer)
     assert.strictEqual(result, renderingInfo.getResultsUrl().replace('__random__', id))
   })
 
@@ -101,8 +111,9 @@ describe('ServerConnector', () => {
     let timestampBefore
     serverConnector._axios.defaults.adapter = async config => {
       const response = {status: 200, config, data: {}, headers: {}, request: {}}
-      if (config.isLongRequest) {
+      if (config.url === 'http://long-request.url') {
         response.status = 202
+        response.headers.location = 'http://polling.url'
         timestampBefore = Date.now()
       } else if (config.isPollingRequest) {
         const timestampAfter = Date.now()
@@ -114,7 +125,7 @@ describe('ServerConnector', () => {
     }
     const delayBeforePolling = [].concat(Array(3).fill(100), Array(3).fill(200), 500)
     await serverConnector._axios.request({
-      isLongRequest: true,
+      url: 'http://long-request.url',
       delayBeforePolling,
     })
     assert.strictEqual(timeouts.length, ANSWER_AFTER)
@@ -122,5 +133,81 @@ describe('ServerConnector', () => {
       const expectedTimeout = delayBeforePolling[Math.min(index, delayBeforePolling.length - 1)]
       assert(timeout >= expectedTimeout && timeout <= expectedTimeout + 10)
     })
+  })
+
+  it('check polling protocol', async () => {
+    const configuration = new Configuration()
+    const serverConnector = new ServerConnector(logger, configuration)
+    const MAX_POLLS_COUNT = 2
+    const RES_DATA = {createdAt: Date.now()}
+    let pollingWasStarted = false
+    let pollsCount = 0
+    let pollingWasFinished = false
+    serverConnector._axios.defaults.adapter = async config => {
+      const response = {status: 200, config, data: {}, headers: {}, request: {}}
+      if (!pollingWasStarted) {
+        response.status = 202
+        response.headers.location = 'http://polling.url'
+        pollingWasStarted = true
+      } else if (config.url === 'http://polling.url') {
+        pollsCount += 1
+        if (pollsCount >= MAX_POLLS_COUNT) {
+          response.status = 201
+          response.headers.location = 'http://finish-polling.url'
+        } else {
+          response.status = 200
+        }
+      } else if (config.url === 'http://finish-polling.url') {
+        response.status = 200
+        response.data = RES_DATA
+        pollingWasFinished = true
+      }
+      return response
+    }
+    const result = await serverConnector._axios.request({
+      url: 'http://long-request.url',
+    })
+
+    assert(pollingWasStarted)
+    assert.strictEqual(pollsCount, MAX_POLLS_COUNT)
+    assert(pollingWasFinished)
+    assert.deepStrictEqual(result.data, RES_DATA)
+  })
+
+  // NOTE: this can be deleted when Eyes server stops being backwards compatible with old SDK's that don't support long running tasks
+  it('sends special request headers for all requests', async () => {
+    const serverConnector = new ServerConnector(logger, new Configuration())
+    serverConnector._axios.defaults.adapter = async config => ({
+      status: 200,
+      config,
+      data: config.headers,
+      headers: {},
+      request: {},
+    })
+
+    const {data} = await serverConnector._axios.request({url: 'http://bla.url'})
+
+    assert.strictEqual(data['Eyes-Expect'], '202+location')
+    assert.ok(data['Eyes-Date'])
+  })
+
+  // NOTE: this can be deleted when Eyes server stops being backwards compatible with old SDK's that don't support long running tasks
+  it("doesn't send special request headers for polling requests", async () => {
+    const serverConnector = new ServerConnector(logger, new Configuration())
+    serverConnector._axios.defaults.adapter = async config => ({
+      status: 202,
+      config,
+      data: config.headers,
+      headers: {},
+      request: {},
+    })
+
+    const {data} = await serverConnector._axios.request({
+      url: 'http://polling.url',
+      isPollingRequest: true,
+    })
+
+    assert.strictEqual(data['Eyes-Expect'], undefined)
+    assert.strictEqual(data['Eyes-Date'], undefined)
   })
 })