+++ packages/side-eyes/src/background/image-strategies/css-stitching/polyfills.js
@@ -1,0 +1,6 @@
+import fs from 'fs'
+import url from 'url'
+
+fs.open = () => {}
+url.URL = URL
+process.hrtime = require('browser-process-hrtime')