+++ packages/side-eyes/src/background/modal-settings/index.js
@@ -1,0 +1,6 @@
+export const incompleteVisualGridSettings = {
+  message: `There are incomplete visual grid settings in the Eyes extension.
+  Playback using the visual grid cannot be completed until the needed settings are in-place.`,
+  cancelLabel: 'Abort playback',
+  confirmLabel: 'Run without the grid',
+}