+++ packages/side-eyes/src/commons/modes.js
@@ -1,0 +1,8 @@
+export default {
+  NORMAL: 'normal',
+  SETUP: 'setup',
+  INVALID: 'invalid',
+  DISCONNECTED: 'disconnected',
+  PLAYBACK: 'playback',
+  RECORD: 'record',
+}