+++ packages/eyes-common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## Unreleased
+
+
+## 3.20.1
+
+- returning batchInfo from toOpenEyesConfiguration
+
+## 3.20.0
+
+- Region supports toPersistedRegions() for VGC
+
+## [3.19.0]
+
+- batchInfo has generating batch id info
+
 ## [3.18.5]
 
 - fix isFeatureFlagOn and isFeatureFlagOff