+++ packages/eyes-images-browser/CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## Unreleased
 
 ## [4.6.54](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.6.53...@applitools/eyes-images-browser@4.6.54) (2019-11-28)