+++ packages/eyes-sdk-core-browser/CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## Unreleased
 
 ## [4.5.52](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-sdk-core-browser@4.5.51...@applitools/eyes-sdk-core-browser@4.5.52) (2019-11-28)